+++ src/entity_handler.rs
@@ -4,3 +4,5 @@ pub mod spawn_manager;
 pub mod factory;
 pub mod enemy_type;
 pub mod recycler;
+pub mod bullet_pool;
+pub mod triangle_assistant;