+++ src/utils.rs
@@ -1,6 +1,4 @@
 pub mod timer;
-pub mod bullet_pool;
 pub mod globals;
 pub mod machine;
 pub mod counter;
-pub mod triangle_assistant;